Fixed inefficient call to doubleValue() spotted by Mikolas into a more efficient longValue().
git-svn-id: svn://svn.forge.objectweb.org/svnroot/sat4j/maven/trunk@881 ba638df5-4473-46d1-82f8-c3ae2a17a6e1

+++ org.sat4j.pb/src/main/java/org/sat4j/pb/orders/VarOrderHeapObjective.java
@@ -75,7 +75,7 @@ public class VarOrderHeapObjective extends VarOrderHeap {
 						p = neg(p);
 					}
 					int var = var(p);
-					activity[var] = c.abs().doubleValue();
+					activity[var] = c.abs().longValue();
 					if (heap.inHeap(var))
 						heap.increase(var);
 					else
